# docs/api/openapi/session-service.yaml
openapi: 3.1.0
info:
  title: Noema Session Service API
  version: 0.1.0
  description: |
    Session Service manages learning session lifecycle, card queuing,
    attempt recording, and session state transitions.

    **Bounded Context:** Learning Sessions
    **Database:** PostgreSQL + Redis
    **Events:** 12 session domain events via Redis Streams

    All mutation responses include `agentHints` for downstream agent consumption.
  contact:
    name: Noema Platform
  license:
    name: Proprietary
servers:
  - url: http://localhost:3003
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ==========================================================================
  # Schemas
  # ==========================================================================
  schemas:
...
    ApiMetadata:
      type: object
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        serviceName:
          type: string
        serviceVersion:
          type: string
        executionTime:
          type: number
...
    ToolDefinition:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        service:
          type: string
        priority:
          type: string
          enum: [P0, P1, P2]
        inputSchema:
          type: object
    ToolExecuteInput:
      type: object
      properties:
        tool:
          type: string
        input:
          type: object
      required: [tool]
...
paths:
...
  /v1/tools:
    get:
      operationId: listTools
      tags: [Tools]
      summary: List available MCP tools (discovery)
      responses:
        "200":
          description: Tool definitions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      tools:
                        type: array
                        items:
                          $ref: "#/components/schemas/ToolDefinition"
                      count:
                        type: integer
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "401":
          description: Unauthorized
  /v1/tools/execute:
    post:
      operationId: executeTool
      tags: [Tools]
      summary: Execute an MCP tool by name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolExecuteInput"
      responses:
        "200":
          description: Tool executed successfully
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
        "404":
          description: Tool not found
        "422":
          description: Tool execution failed

  # ---------- Health ----------
...